refactor: moves sync state logic out of boardEditor
