+++ src/utils/board/editor/boardEditor.ts
@@ -5,16 +5,15 @@ import { EMPTY_POSITION } from '../board.constants';
 import initialBoard from '../board.init';
 import {
   BoardFile,
-  BoardKingState,
   BoardPosition,
   BoardRank,
-  BoardSquare,
   BoardSubstate,
   Coordinate,
 } from '../board.types';
 import { pieceObjectToString, pieceStringToObject } from '../../pieces';
 import { PieceString } from '../../pieces.types';
-import { DevError } from '../../errors';
+import { ConcisePosition } from './boardEditor.types';
+import { synchronizeKingState } from './syncState';
 
 export function createBoard(board: {
   position?: ConcisePosition;
@@ -38,44 +37,6 @@ function synchronizeState(position?: ConcisePosition, state?: BoardSubstate) {
   return assign({}, syncedState, { king: kingState });
 }
 
-function synchronizeKingState(
-  position?: ConcisePosition,
-  state?: BoardSubstate
-): BoardKingState {
-  const priorKingState: BoardKingState = produce(
-    initialBoard.state.king,
-    (draft) => {
-      if (state?.king?.w?.square)
-        draft.w.square = state.king.w.square as BoardSquare;
-      if (state?.king?.b?.square)
-        draft.b.square = state.king.b.square as BoardSquare;
-    }
-  );
-
-  return produce(priorKingState, (draft) => {
-    if (!position) return;
-    if (Array.isArray(position.wk) && position.wk.length !== 1)
-      throw new DevError('Only one white king.');
-    if (Array.isArray(position.bk) && position.bk.length !== 1)
-      throw new DevError('Only one black king.');
-
-    const whiteKingCoordinate = Array.isArray(position.wk)
-      ? position.wk[0]
-      : position.wk;
-    const blackKingCoordinate = Array.isArray(position.bk)
-      ? position.bk[0]
-      : position.bk;
-
-    if (position.wk)
-      draft.w.square = coordinateToSquare(whiteKingCoordinate as Coordinate);
-    if (position.bk)
-      draft.b.square = coordinateToSquare(blackKingCoordinate as Coordinate);
-  });
-}
-
-export type ConcisePosition = {
-  [pieceString in PieceString]?: Coordinate[] | Coordinate;
-};
 export function createFromConcisePosition(concisePosition: ConcisePosition) {
   return produce(EMPTY_POSITION, (draft) => {
     for (const pieceString in concisePosition) {
