+++ src/utils/board/editor/boardEditor.types.ts
@@ -1,0 +1,6 @@
+import { PieceString } from '../../pieces.types';
+import { Coordinate } from '../board.types';
+
+export type ConcisePosition = {
+  [pieceString in PieceString]?: Coordinate[] | Coordinate;
+};
